requirements.txt
discord.py>=2.3.0
gspread>=6.0.0
google-auth>=2.0.0
matplotlib>=3.8.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
